misc/ai-spam-email-monitor/requirements.txt
google-api-python-client==2.118.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.3.2